+++ routes/r_screening.js
@@ -19,26 +19,31 @@ router.get('/view', authContoller.isLoggedIn, (req, res) => {
 
     //if there is request from user with jwt token
     if (req.user) {
+        if (req.user.role === "Patient") {
 
-        db.query('SELECT * FROM screening WHERE ic = ?', [req.user.ic], (err, rows) => {
-            //when done with connection
+            db.query('SELECT * FROM screening WHERE ic = ?', [req.user.ic], (err, rows) => {
+                //when done with connection
 
-            if (!err) { //if not error
-                let removedScreening = req.query.removed; //if any screening is deleted, set alert 
-                db.query('SELECT * FROM patientdetails WHERE ic = ?', [req.user.ic], (error, row) => {
-                    if (!error) {
-                        res.render('v_p_screening', { user: req.user, rows, removedScreening: removedScreening, assignedTo: row[0].assignedTo });
-                    } else {
-                        console.log(error);
-                    }
-                })
+                if (!err) { //if not error
+                    let removedScreening = req.query.removed; //if any screening is deleted, set alert 
+                    db.query('SELECT * FROM patientdetails WHERE ic = ?', [req.user.ic], (error, row) => {
+                        if (!error) {
+                            res.render('v_p_screening', { user: req.user, rows, removedScreening: removedScreening, assignedTo: row[0].assignedTo });
+                        } else {
+                            console.log(error);
+                        }
+                    })
 
 
-            } else {
-                console.log(err);
-            }
-            console.log(rows);
-        })
+                } else {
+                    console.log(err);
+                }
+                console.log(rows);
+            })
+
+        } else {
+            res.status(404).send('Not Found');
+        }
 
     } else {
         res.redirect('/login');
@@ -49,27 +54,32 @@ router.get('/view', authContoller.isLoggedIn, (req, res) => {
 router.get('/add', authContoller.isLoggedIn, (req, res) => {
     //if there is request from user with jwt token
     if (req.user) {
+        if (req.user.role === "Patient") {
 
-        db.query('SELECT * FROM screening WHERE ic = ?', [req.user.ic], (err, rows) => {
-            //when done with connection
+            db.query('SELECT * FROM screening WHERE ic = ?', [req.user.ic], (err, rows) => {
+                //when done with connection
 
-            if (!err) { //if not error
+                if (!err) { //if not error
 
-                db.query('SELECT * FROM patientdetails WHERE ic = ?', [req.user.ic], (error, row) => {
-                    if (!error) {
-                        res.render('v_p_screening_add', { user: req.user, rows, assignedTo: row[0].assignedTo });
-                    } else {
-                        console.log(error);
-                    }
-                })
+                    db.query('SELECT * FROM patientdetails WHERE ic = ?', [req.user.ic], (error, row) => {
+                        if (!error) {
+                            res.render('v_p_screening_add', { user: req.user, rows, assignedTo: row[0].assignedTo });
+                        } else {
+                            console.log(error);
+                        }
+                    })
 
 
-            } else {
-                console.log(err);
-            }
-            console.log(rows);
+                } else {
+                    console.log(err);
+                }
+                console.log(rows);
 
-        })
+            })
+
+        } else {
+            res.status(404).send('Not Found');
+        }
 
     } else {
         res.redirect('/login');
@@ -78,38 +88,43 @@ router.get('/add', authContoller.isLoggedIn, (req, res) => {
 
 router.get('/update/:id', authContoller.isLoggedIn, (req, res) => {
     if (req.user) {
+        if (req.user.role === "Patient") {
 
-        //if there is request from user with jwt token
-        db.query('SELECT * FROM screening WHERE ic = ?', [req.user.ic], (err, row) => {
-            //when done with connection
+            //if there is request from user with jwt token
+            db.query('SELECT * FROM screening WHERE ic = ?', [req.user.ic], (err, row) => {
+                //when done with connection
 
-            if (!err) { //if not error
+                if (!err) { //if not error
 
-                db.query('SELECT * FROM screening WHERE id = ? AND ic = ?', [req.params.id, req.user.ic], (err, rows) => {
-                    //when done with connection
+                    db.query('SELECT * FROM screening WHERE id = ? AND ic = ?', [req.params.id, req.user.ic], (err, rows) => {
+                        //when done with connection
 
-                    if (!err) { //if not error
+                        if (!err) { //if not error
 
-                        db.query('SELECT * FROM patientdetails WHERE ic = ?', [req.user.ic], (error, row) => {
-                            if (!error) {
-                                res.render('v_p_screening_edit', { user: req.user, rows, update: 'update', assignedTo: row[0].assignedTo });
-                            } else {
-                                console.log(error);
-                            }
-                        })
+                            db.query('SELECT * FROM patientdetails WHERE ic = ?', [req.user.ic], (error, row) => {
+                                if (!error) {
+                                    res.render('v_p_screening_edit', { user: req.user, rows, update: 'update', assignedTo: row[0].assignedTo });
+                                } else {
+                                    console.log(error);
+                                }
+                            })
 
-                    } else {
-                        console.log(err);
-                    }
-                    console.log(rows);
-                })
+                        } else {
+                            console.log(err);
+                        }
+                        console.log(rows);
+                    })
 
-            } else {
-                console.log(err);
-            }
-            console.log(row);
+                } else {
+                    console.log(err);
+                }
+                console.log(row);
 
-        })
+            })
+
+        } else {
+            res.status(404).send('Not Found');
+        }
 
     } else {
         res.redirect('/login');
@@ -119,34 +134,39 @@ router.get('/update/:id', authContoller.isLoggedIn, (req, res) => {
 router.get('/:id', authContoller.isLoggedIn, (req, res) => {
     //if there is request from user with jwt token
     if (req.user) {
+        if (req.user.role === "Patient") {
 
-        db.query('SELECT * FROM screening WHERE ic = ?', [req.user.ic], (err, row) => {
-            //when done with connection
+            db.query('SELECT * FROM screening WHERE ic = ?', [req.user.ic], (err, row) => {
+                //when done with connection
 
-            if (!err) { //if not error
+                if (!err) { //if not error
 
-                if (req.params.id) {
-                    db.query('DELETE FROM screening WHERE id = ? AND ic = ?', [req.params.id, req.user.ic], (err, rows) => {
-                        //when done with connection
+                    if (req.params.id) {
+                        db.query('DELETE FROM screening WHERE id = ? AND ic = ?', [req.params.id, req.user.ic], (err, rows) => {
+                            //when done with connection
 
-                        if (!err) { //if not error
-                            let removedScreening = encodeURIComponent('Screening Successfully Removed');
-                            res.redirect('/screening/view?removed=' + removedScreening); //no need render just redirect to same page of current page dislaying
-                            //res.redirect('/diet/view');
-                        } else {
-                            console.log(err);
-                        }
-                        console.log(rows);
-                    })
+                            if (!err) { //if not error
+                                let removedScreening = encodeURIComponent('Screening Successfully Removed');
+                                res.redirect('/screening/view?removed=' + removedScreening); //no need render just redirect to same page of current page dislaying
+                                //res.redirect('/diet/view');
+                            } else {
+                                console.log(err);
+                            }
+                            console.log(rows);
+                        })
+                    }
+
+
+                } else {
+                    console.log(err);
                 }
+                console.log(row);
 
+            })
 
-            } else {
-                console.log(err);
-            }
-            console.log(row);
-
-        })
+        } else {
+            res.status(404).send('Not Found');
+        }
 
     } else {
         res.redirect('/login');
@@ -156,42 +176,47 @@ router.get('/:id', authContoller.isLoggedIn, (req, res) => {
 router.get('/display/:id', authContoller.isLoggedIn, (req, res) => {
     //if there is request from user with jwt token
     if (req.user) {
+        if (req.user.role === "Patient") {
 
-        db.query('SELECT * FROM screening', (err, row) => {
-            //when done with connection
+            db.query('SELECT * FROM screening', (err, row) => {
+                //when done with connection
 
-            if (!err) { //if not error
+                if (!err) { //if not error
 
-                if (req.params.id) {
+                    if (req.params.id) {
 
-                    db.query('SELECT * FROM screening WHERE id = ? AND ic = ?', [req.params.id, req.user.ic], (err, rows) => {
-                        //when done with connection
+                        db.query('SELECT * FROM screening WHERE id = ? AND ic = ?', [req.params.id, req.user.ic], (err, rows) => {
+                            //when done with connection
 
-                        if (!err) { //if not error
+                            if (!err) { //if not error
 
-                            db.query('SELECT * FROM patientdetails WHERE ic = ?', [req.user.ic], (error, row) => {
-                                if (!error) {
-                                    res.render('v_p_screening_display', { user: req.user, rows, alert: 'Your Selected Screening Displayed Below', assignedTo: row[0].assignedTo });
-                                } else {
-                                    console.log(error);
-                                }
-                            })
+                                db.query('SELECT * FROM patientdetails WHERE ic = ?', [req.user.ic], (error, row) => {
+                                    if (!error) {
+                                        res.render('v_p_screening_display', { user: req.user, rows, alert: 'Your Selected Screening Displayed Below', assignedTo: row[0].assignedTo });
+                                    } else {
+                                        console.log(error);
+                                    }
+                                })
 
 
-                        } else {
-                            console.log(err);
-                        }
-                        console.log(rows);
-                    })
+                            } else {
+                                console.log(err);
+                            }
+                            console.log(rows);
+                        })
+                    }
+
+
+                } else {
+                    console.log(err);
                 }
+                console.log(row);
 
+            })
 
-            } else {
-                console.log(err);
-            }
-            console.log(row);
-
-        })
+        } else {
+            res.status(404).send('Not Found');
+        }
 
     } else {
         res.redirect('/login');
